requirements.txt
cryptography==3.4.7
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.15.1